requirements.txt
fastapi
uvicorn
python-dotenv
databricks-sdk==0.65.0
psycopg2-binary
pydantic
geoalchemy2
shapely
geojson